Add import from MapIt rake task
Import local authorities from MapIt.  We want the name, GSS code and SNAC (also
 known as ONS) code for each.  We derive the authority tier from the area type.

Slugs and home page URLs will be added in subsequent import tasks.

Although SNAC codes and GSS codes are both assumed to be immutable, in reality
we want to have some flexibility with this.  We've changed SNAC codes before
(e.g. with Northern Ireland LAs) so want to retain the ability to do this.