Add LocalAuthority model
`name`, `gss`, `snac` and `tier` are validated for presence as these are
imported or derived from MapIt in the initial data import and are all present
at this time.

gss and snac codes are indexed and unique as they are IDs for authorities and
are likely to be used in queries.

`snac` is referred to as `ons` in both MapIt and Frontend, but the team voted
to use `snac` as they are both "ons" codes.

`tier` may be one of county, district or unitary