Create CsvDownloader class
We create a class for downloading CSV files which we can reuse in other classes to import CSV data. We use the Net::HTTP module and match against Net::HTTPOK for success. We raise a custom error if there is no match, partially in order to get around having to rescue the large number of Net::HTTP errors