Create InteractionsImporter class
We create the class to import Interactions from a preset URL. We use the `CsvDownloader` class to import the CSV data, using Dependency Injection (for easier testing). We log errors if the import fails. For each imported CSV row, the importer either creates a new record or updates an existing record.