Create ServiceInteractionsImporter class
We create the class to import ServiceInteractions from a preset URL. We use the `CsvDownloader` class to import the CSV data, using Dependency Injection (for easier testing). After the import, we check that the associated Service and Interaction have already been created, raising errors if not.
We log errors if the import fails. For each imported CSV row, the importer either creates a new record or updates an existing record.