Add detailed summary logging for ServiceInteractions
When importing ServiceInteractions, we have identified a number of things that can wrong, such as missing fields in the CSV or linking to Services or Interactions that that do not exist.
Previously we only log when something goes wrong for each entry, which is hard to read in the logs. This commit adds a convenient summary into the log at the end of the import.