Use CSV header converters
In the Importer classes (Service, Interaction, ServiceInteraction), we previously converted some field names to more readable names (such as "Mapped identifier" -> :lgil_code). This forced another level of indirection for imported CSV rows, having to call the `parsed_hash` method on a row first before it could get passed to the `create_or_update_method`.

This commit makes use of the `CSV` library's support for [converters](http://ruby-doc.org/stdlib-2.0.0/libdoc/csv/rdoc/CSV.html#method-i-header_convert), so we can pass a hash of conversions to the CsvDownloader class, and the conversions are applied as the CSV rows are imported. This helps to declutter the Importer classes, makes the CsvDownloader responsible for the conversions and makes the tests more readable.