Add NonBlankUrlValidator
We took this from transition where it is used to validate that various
urls in that system are valid.  It's a bit more robust than a simple
regexp check, and allows us to do more checking if we need it (e.g.
we could restrict to http(s) only by checking the scheme after parsing.

We make the addressable gem an explicit dependency because the validator
uses it.

Finally, we make the validation for LocalAuthority#homepage_url use this
validator.