Display links for an LA/service/interaction in the UI
We display the link as a clickable anchor tag if it exists, or as the
string 'n/a' if not.  Also included is a simple table row matcher for
our features that passes if any tr tag on the page (or scoped selector)
contains the exact array of text supplied as cells.  A future version
could be upgraded to allow using other rspec or capybara matchers instead
of exact text matches.