Add ServiceInteraction#find_by_lgsl_and_lgil
This would normally be a scope, but the intention is that it returns a
single record (or nil) and scopes always return an ActiveRecord::Relation
(e.g. an array of results) even if you put .first on the end.  This would
mean every call-site of the scoped method would have to add `.first` to
the end and that's hardly ideal.  It does mean you can't chain other
scopes on to the end of the call, but it will respect other scopes chained
to the start of the call (like .find_by).