Improve CsvDownloader handling of encoding errors
We needed to improve the encoding handling as the CSV used to import links
has latin1 / windows 1252 characters in it that unless handled correctly
result in postgres throwing errors about invalid utf8 sequences.

The new implementation of CsvDownloader#downloaded_csv is essentially
the same as the one for [LocalAuthorityDataImporter.fetch_http_to_file in
Publisher][1].  The difference here is that we use the block form of
Tempfile.create so that the tempfile is closed and deleted once we're done
with it.  We also allow for specifying which encoding the downloaded CSV
file is in so that it can be converted to utf8 correctly (we default to
windows-1252 like Publisher does though).

To make using the block form of Tempfile.create work we change
CsvDownloader#download to take a block and yield the csv parser. We also
introduce CsvDownloader#each_row which wraps CsvDownloader#download to
iterate over the rows in the CSV.  This is what every user of CsvDownloader
did with the #download anyway so we update them all to use this new method.

Little bit unhappy with the specs as changing from using
CsvDownloader#download to CsvDownloader#each_row meant a lot of churn in
the specs because of stubbing.  And yet, it still feels like stubbing is
the correct thing to do here as we're not that interested in how
CsvDownloader really works, just that we get yielded some rows from it.

[1]: https://github.com/alphagov/publisher/blob/master/lib/local_authority_data_importer.rb#L48-L64)