Ignore known problems while importing links
We know that the links csv from local direct gov has some rows that we
don't care about so we ignore those explicitly rather than mixing them
in with the unexpected errors.

* rows with a blank SNAC - these are rows for authorities that aren't
  local authorities, mostly they deal with issuing licenses, and we
  don't care about them
* rows with an 'x' url - these indicate that the authority provides the
  service but the url is special somehow.  It's mostly licenses and
  again we don't care
* rows for the old NI councils - these councils don't exist anymore
  and we don't care what their links were

In all cases we ignore the rows and increment the ignored_rows_count
variable rather than raising a more specific error.  To test this we
expose the counts as readers on the importer.