Create LinkChecker class
The LinkChecker class returns the status (HTTP status code or custom error) and time that the link was checked for each link. We use Faraday to make HTTP GET requests. We attempt to explicitly catch the most common errors, based on trial runs. We catch all errors just to make sure that the process finishes as it will likely take 20+ hours to complete for 36000+ links.

Note that we chose to use `headers: { accept_encoding: 'none' }` option when initialising the Faraday connection to get around this issue -> https://github.com/lostisland/faraday/issues/120. Certain URLs were causing the `Zlib::DataError: unknown compression method` error so we used the suggested workaround. The cause is suggested to be a combination of a Ruby bug in net/http and the compression that the server uses to send the response.