Add disabled_service factory
Make the default Service Factory be enabled, as we most likely will be testing Links for enabled Services more often, thus reducing the extra effort of setting a Service to be enabled for every test.
We update the one test that uses disabled Services. The second test was not actually invoking the import task, so we update it.