Add Icinga passive check
We changed the distributed_lock to take callbacks for what to do when
the lock is obtained and when it is not obtained. This follows the
pattern in the Publisher script where if the lock is not obtained we
tell Icinga it was a success but we did not actually run the task.

We now communicate with Icinga using a new notify_passive_check script
which was added to Puppet.