Alert Icinga if records are missing from imports
We want to know if we've previously imported services, interactions or local
authorities which subsequently disappear from the CSV.  We can't delete these
automatically as we'll have to delete a load of dependent `ServiceInteraction`s
and potentially useful links.  This would also have a knock on for any
published `local_transaction` pages we have published.

Because we don't know at this point what this would mean, we'll just alert and
decide what to do later.

Trello card: https://trello.com/c/KuN7jtEr/394-imports-should-delete-things-sensibly-5

The code to determine whether records have been removed from the source CSV is
common across all the things that we want to check.  It doesn't make sense to
maintain this in multiple places, so we'll extract it here.

The key attribute of each object is different (e.g. lgsl for services, lgil for
interactions, gss for local authorities) so we need to be able to pass it in
to the comparison logic.