Alert Icinga when blank homepage URLs are present
We rely on homepage URLs being present as they are the ultimate fall back for
local_transaction URLs if no more specific link exists.  It is important that
we have something present. The link checker should run after the import to tell
us if the URLs are incorrect.

We're using the DistributedLock as seen in other imports in this project.

Also removed unnecessary fields from FactoryGirl construction as we have
sequences set up to cope with things like gss codes now.