Correct enabled service checker
We were incorrectly expecting the enabled service CSV file to provide us with
numeric values (when they are strings).  This threw out the comparator which
enables/disables the services we're interested in resulting in us disabling all
of the services.

We've fixed the tests to be more reflective of reality and now compare strings
instead.