Refactor to replace helper with presenters
We adopt a simple presenter pattern to reduce usage of helper methods. Using Presenters allows for more object-oriented design and explicit assignment of responsibility to appropriate classes. They also help to handle the relationship between links and interactions, which before were a bit awkward to deal with as they relied on a  dictionary lookup.

We recognise that there is scope to extract URL-related behaviour from the LocalAuthority Class regarding its homepage. This will be addressed in a future PR.