Update lozenge styling and fix Link Status display
This commit adds extra space to the right of the lozenges and displays
the Link Status as grey when there is no link.

We had to move both opening and closing <span> elements onto one
line to get the correct spacing. The following link displays the effect in
detail: http://stackoverflow.com/a/25879479