Refactor importer classes to return response object
We wanted to separate the calls to Icinga from inside the ImportComparer class, as we want to report more complex error states which would be handled outside the ImportComparer.
Instead we opted to refactor all Importer classes to return a response object, which would collect any number of errors that could happen during an import. We use this response object in our rake tasks to report the success or failure and summary of errors to Icinga.