Update local authorities importer to import council hierarchy from mapit
We need this hierarchy to support a new local authority API that will return both the district and the county when asked for the district information.  The hierarchy information is modelled in mapit via the parent_area attribute, which we added in https://trello.com/c/XJ4pDHxG/433-1b-of-3-add-parent-child-relationships-to-mapit-3

This commit also updates the importer to use the new response object and communicate with Icinga in the rake task, like we did for all others in the previous commit.