Return LocalAuthority details for missing ServiceInteraction
When we request a LGSL and LGIL combination from the local links API which does not exist, we should still return the LocalAuthority details. Previously only a 404 and and an empty hash was returned, which meant Frontend was serving its "Sorry, we are experiencing technical difficulties" page. Instead Frontend should show the homepage for the given Service and LocalAuthority.

This fix was prompted by visiting '/report-stray-dog' which is LGSL 432 and LGIL 0. This is a ServiceInteraction that we do not support in LocalLinksManager.