Increase unicorn workers to 4
Local links manager is now serving production traffic via its API.  We expect
the traffic to this API to increase when we bring the `find-your-local-council`
service online.  To ensure that we can manage this extra traffic, we're
increasing the number of unicorn worker processes to 4.

This is inline with Imminence which fulfills a similar role for `place` formats
as Local Links Manager does for `local_transactions`.