Manually unlock locked rake tasks
There have been instances where rake tasks have not completed for
for some reason leaving the Distributed Lock for that task in
place. It is then necessary to wait for the lifetime of the lock
to end before being able to run task again.

This rake task adds the ability for the lock to be removed for the
given task passed in.

`Rails.env` was removed from the key as it was felt to be
superfluous since all rake tasks run in the same environment.

We're not using `Services.redis.unlock` here because that doesn't
let us specify the `owner` param.

[Trello card](https://trello.com/c/OtTKMe3H/465-enable-unlocking-of-scheduled-tasks-2-pair-with-katie-pls)