Make most importers provide import_summary with counts
With the exception of the enabled_service_checker which is quite different
we can make all our importers provide an import summary based on counting
what they do.  We follow the pattern established in service tiers importer
which was the last written and already did the most counting.

There's now a lot of duplication in these importers, so our next step will
be to look for ways to extract this shared behaviour and reduce the
duplication.