Extract summariser and processor classes for importers
The summariser deals with counting the things that happen during the
import and providing a summary at the end.  This lets us remove all
the `xxx_count` instance variables, `#import_summary` and
`#counting_errors` as these are all now part of the summariser.  I'm a
bit unsure about this implementation as it's quite verbose and has static
counters.  However, I'm not sure that a more dynamic version where you
can tell it what things to count would be much better than a hash.  At
least the current form will break if we try to count something that
doesn't exist (like a typo).

The processor deals with orchestrating the import process and triggers
callback methods on the importer class itself to deal with importing a
single item.  This lets us remove all the boilerplate about setting up
a `Response` instance, writing out a summary, iterating over the data
source, and dealing with exception handling.  We remove the special
handling of the `CsvDownloader::Error` to keep the processor agnostic
to the type of data being iterated over.