Convert EnabledServiceChecker to use new Processor and Summariser
This "importer" is the most different to the others, and we hadn't tried
in previous commits to make it similar by adding the `#counting_errors`,
`#with_each_row`, or `#import_summary` methods.  However, it doesn't take
much to make it work within the framework provided by Processor and
Summariser.  The only change is that we allow Summariser to take a new
"extra summary" for the part of the log output from EnabledServiceChecker
that isn't a simple set of counts.