Use correct variables in Summariser#import_summary
While building we flipped from `data_source_xxx` to `import_source_xxx`
as the names for the internal variables to record the name and count
values for the imported data.  We changed the initializer and increment
methods, but forgot to change the reference in the `#import_summary`
methods.  Happily, ruby will treat all instance variable references as
`nil`, so nothing breaks but we do get odd log output.