Strip white space around interaction URLs
We have ~240 URLs that fail link checking because of `Invalid URI`.  A short
investigation revealed that ~220 of these have trailing white space on the URL
which will cause URI.parse to throw `URI::InvalidURIError` prior to the link
being checked.  (They may have other problems as well)

We already have code to strip white space in [links_controller][1] and
[local_authorities_controller][2], but this is only called when a URL is edited.

`homepage_urls` are not affected by this problem.

[1]: https://github.com/alphagov/local-links-manager/blob/16c4f0ae9f48f72a981a5142b1e5619fdc416fd4/app/controllers/links_controller.rb#L19
[2]: https://github.com/alphagov/local-links-manager/blob/16c4f0ae9f48f72a981a5142b1e5619fdc416fd4/app/controllers/local_authorities_controller.rb#L12