Update LinkStatusUpdater to also update LocalAuthority broken link count
When the status of a Link changes, we should update the Local Authority.
There's no need to do anything particularly clever - it just iterates
over the full list of LAs after the link-checker completes - since the
time-complexity of this routine is entirely bound to the network and
response times of the remote services.