Remove the after_save callback on Link#status
ActiveRecord callbacks are a bad idea for a number of reasons. I came
across one of these reasons while trying to test the interaction between
`Links#update` and `LocalAuthority#update_count_of_broken_links` - in
order to get the test working I had to work-around the lifecycle
callback, which hid the true intention of the test and added a lot of
clutter.