Update the specific council service page based on the prototype
We've changed the routes to not be
`local-authorities/local_authority_name/services/service_name/interactions`
but instead `local-authorities/local_authority_name/interaction_name` as
per the prototype layout. Fixed all the failing tests as a result of this.

We also took out the LGIL code, and replaced "LGSL" with "Service code" as
a description. We also removed the homepage URL displaying.

Lastly, we removed the ability for people to add new links, pending further
user research.