Use uniq over distinct
It seems that using `.uniq` on the array of urls is more than an order of
magnitude quicker than using `.distinct` here.  (~50ms vs ~850ms on prod)

The number of URLs is likely to stay fairly static (40,000ish).

It doesn't matter very much relative to the length of the task, but since I
spotted it, I thought it was worth doing.