Add Google Analytics download code
We want to get analytics for external links that are clicked on within local
transactions.  The intention is that we will use these to rank links in local
links manager.  More clicks means that the link is more important and should
be prioritised.

Running the GA code locally can be done by creating a file at
`config/local_env.yml`.  This should contain values for `GOOGLE_PRIVATE_KEY`,
`GOOGLE_CLIENT_EMAIL` and `GOOGLE_ANALYTICS_GOVUK_VIEW_ID`.  You can either
grab an example file from me, or construct your own with other credentials.
These values are set as environment variables in our non-Dev environments.

The import task adds the analytics count to each link. Not all links will
appear in GA.  In addition, it's possible that our query will return things
that we're not interested in.
When we're confident that we've had a successful import, we want to reset the
count on the other links to zero to ensure that we're only counting the last
week.

We don't want to use the local link to do the lookup, as it's possible that
this has changed over the week.  We're interested in the popularity of a
ServiceInteraction for a Local Authority, not any particular URL.

The GA client is a similar approach to that used by content-performance-manager