Update Link model callbacks
Now that this app is using the Link Checker API, more link check information is
being persisted in the new `link_errors` and `link_warnings` attributes.  This
updates the callbacks on the Link model so that these attributes are set and
updated appropriately when a link is created or updated.  We also take the
opportunity to rename some methods names that weren't very clear.