Import titles and slugs of local transactions
We're going to present the title of local transactions above the links which
will be shown on that transaction.  This will help administrators to choose a
more accurate link to match the content on the GOV.UK page.

The slug is to be used to provide a link to the local transaction page so that
an adminstrator can navigate directly to the page and check the body of the
content if necessary.

The slug and the title are both assigned to a local transaction page which has
a combination of LGIL and LGSL code.  This maps closely to the concept of a
service interaction which is why we've added them to that model.

The slug is also going to be used to marry up Google analytics for each page/

We're also tracking which service interactions are actually live on GOV.UK.  We
have a lot of links which correspond to LGSL/LGIL combinations that aren't
ever used.  By using this import we can check which ones are used for GOV.UK,
and only show those ones to administrators.  We can also delete the unused ones
as they will not be curated.

We should track if we consider a service to be enabled, and it then disappears
from the import.  This could be down to user error, or the retiring of a local
transaction.  Either way, we want to know about it - which is why an error will
be raised.

The fallback code will be removed once all local transactions have an LGIL code
assigned.