Only reset analytics count on necessary links
We don't want to reset the analytics count on every link that is not in the GA
import because most of them are already at zero.  Instead, we'll make a note
of which links have a non-zero count beforehand, and just reset those that
haven't been updated in the import.