Return 404 if resources can't be found
At present, these two controllers may load nil resources which are then passed
onto the view resulting in 500 errors.  This happens a bit at the moment
because of the way we add missing links.

These will now return a 404 instead if you make a mistake, which is both more
correct and helpful to the end user.