Change link checking strategy
If we submit all the links for a local authority at once, we have noticed that
several authorities fail link checking for lots of their links.  I guess that
this is because they haven't warmed up sufficiently at 2AM when the link
checker runs.  We run 4 threads on the sidekiq worker at present and it's
possible that we are causing our own bottleneck.

I propose a different strategy where we check links by service instead.  This
spaces out the requests to local authority websites a bit which allows them
extra time to warm up.

I've also changed the order so that the least significant (by GA click count)
is requested first.  I'm hoping that this will contribute to a more accurate
link checker status for the high value links in local links manager which is
what we're after.

We check the homepages last of all as it's most important that they are correct.

I've also added a rake task allowing us to recheck a whole council, as
sometimes they are down for maintenance overnight. There is nothing that our
admins can do, so we should be able to easily recheck.