Reduce number of updates when a batch is received
Currently we update service and local authority broken link counts every time
we process a URL in a batch receipt.  This probably means that we update the
count for either a local authority or a service many times in a single batch
receipt.

(We tend to submit a lot of URLs belonging to either the same service
 or the same local authority in a batch)

By assembling a list of the services and local authorities that need updating
and doing it once at the end of the batch, we hope to reduce the time taken
for the request to complete.