Allow empty URLs
We want to be able to track which service/interaction/council combinations are
missing a link.  We also want to be able to track the same analytics against
these missing links as we do those that we have.  (This is so that we can
prioritise the fixing of these missing links in the same way that we do the
broken ones.)

So that we can do this, we'll allow the unsetting of the URL field of the link
and treat nil URLs as missing ones.