Remove interactions controller and related items
This functionality allows us to list all the links provided for a service by a
local authority.  We're going to remove it.

There isn't a need for this any more because we no longer use the fallback
logic where if we didn't have a link for the specific LGSL/LGIL combination
we'd check use one from another LGIL.  (This didn't make a lot of sense
particularly because link quality is more variable among little used links)