Update googleauth gem to fixed auth version
This will update the googleauth gem to version 0.5.3 where the usage of
`Google::Auth.get_application_default(scopes)`
has been fixed. In which case we don't have to go digging in the google API to generate valid credentials like we've done here: https://github.com/alphagov/feedback/blob/ebe2a359be75ebf339dafbb405cdbc

In order for the `google-api-client` gem to work with the new version of `googleauth`, we've had to bump the version up from 0.10.0 to 0.17.0 (which is currently the latest).

This in turn has led to a cascade of other gems having to be **conservatively** updated in order to work with either `googleauth` or `google-api-client`.