Updates documentation for GA uploads via API
This will update the documentation to reflect what we've discovered about using the Google Management API.

We discovered that you can't use multiple service accounts when working with the GM API, because it expects some credentials to be only set once.
The are the `GOOGLE_PRIVATE_KEY` and the `GOOGLE_CLIENT_EMAIL`.

These cannot be set for multiple service accounts, and the API will look for them when attempting to authenticate.

We've added documentation with a workaround for this constraint, which is basically recommending to re-use the same service account but with multiple scoped permissions that will allow you to do your tasks.

For example if you are trying to upload a file to GA, but you're also trying to write to Google Spreadsheets you can use the same `GOOGLE_PRIVATE_KEY` and `GOOGLE_CLIENT_EMAIL`,
but at the same time set the permissions on the service account as:

SCOPES = [`https://www.googleapis.com/auth/analytics.edit`, `https://www.googleapis.com/auth/spreadsheets`]