Schedule deleting uploaded files from GA
For: https://trello.com/c/NYNZzVyW/315-schedule-deleting-files-from-ga

Because we are now uploading a daily list of bad links to GA, we are going to reach the quota limit if we don't start deleting these files automatically.

Deleting the files after they've been processed does not get rid of the data that was introduced through them.

For reference and as a sanity check that the daily upload task is still working, we're going to delete all files but the last 2.

This job will run daily, before we upload a new file.