Use govuk_app_config to configure unicorn
We move `worker_processes 4` to be configured by an env var controlled by
puppet.  We leave `working_directory` bit because it isn't handled by
govuk_app_config, and appears to be used by several other apps with a
unicorn.rb (it could be a cargo-cult, but let's not chance it yet).