Resolve deprecation warnings
This resolves the deprecation warning:

> The success? predicate is deprecated and will be removed in Rails 6.0. Please use successful? as provided by Rack::Response::Helpers.

There is a [fix for this that has been merged](https://github.com/rspec/rspec-rails/pull/1951)
for rspec-rails, but is [not yet released](https://github.com/rspec/rspec-rails/blob/678b313cf3cc3bdb2c59149ee3e290fd590460d1/Changelog.md).
We could revert it back afterwards if we're worried, but we expect this to be
a `200` response at the moment, not a `202` or anything so this is fine I think.