Send string instead of boolean to icinga check
The `code` argument passed to `self.icinga_check` is a boolean value so we
can't call `shellescape` on it. This was causing the `check-links` rake task to
break.

Initially I removed `shellscape` from `code` in the `self.icing_check` method,
however Brakeman is flagging `code` as a potential command injection security
warning because it is a boolean value. The only apparent use of `code` is to be
interpolated in a string which is used to notify Icinga, so it seems reasonable
to send it as a string instead.